requirements.txt
python-dotenv~=1.0.1
pyinstaller~=6.7.0
PyQt6~=6.7.0
pyqtdarktheme~=2.1.0
